+++ packages/core/src/actions.ts
@@ -41,7 +41,8 @@ import {
   EventFrom,
   AnyActorRef,
   PredictableActionArgumentsExec,
-  RaiseActionOptions
+  RaiseActionOptions,
+  BaseActionObject
 } from './types';
 import * as actionTypes from './actionTypes';
 import {
@@ -90,35 +91,35 @@ export function toActionObject<TContext, TEvent extends EventObject>(
       actionObject = {
         type: action,
         exec
-      };
+      } as any;
     } else if (exec) {
       actionObject = exec;
     } else {
-      actionObject = { type: action, exec: undefined };
+      actionObject = { type: action, exec: undefined } as any;
     }
   } else if (isFunction(action)) {
     actionObject = {
       // Convert action to string if unnamed
       type: action.name || action.toString(),
       exec: action
-    };
+    } as any;
   } else {
-    const exec = getActionFunction(action.type, actionFunctionMap);
+    const exec = getActionFunction((action as any).type, actionFunctionMap);
     if (isFunction(exec)) {
       actionObject = {
-        ...action,
+        ...(action as any),
         exec
       };
     } else if (exec) {
-      const actionType = exec.type || action.type;
+      const actionType = (exec as any).type || (action as any).type;
 
       actionObject = {
-        ...exec,
-        ...action,
+        ...(exec as any),
+        ...(action as any),
         type: actionType
       } as ActionObject<TContext, TEvent>;
     } else {
-      actionObject = action as ActionObject<TContext, TEvent>;
+      actionObject = action as any;
     }
   }
   return actionObject;
@@ -145,10 +146,10 @@ export function toActivityDefinition<TContext, TEvent extends EventObject>(
   const actionObject = toActionObject(action);
 
   return {
-    id: isString(action) ? action : actionObject.id,
+    id: isString(action) ? action : (actionObject as any).id,
     ...actionObject,
     type: actionObject.type
-  };
+  } as any;
 }
 
 /**
@@ -166,7 +167,7 @@ export function raise<TContext, TEvent extends EventObject>(
     event: typeof event === 'function' ? event : toEventObject<TEvent>(event),
     delay: options ? options.delay : undefined,
     id: options?.id
-  };
+  } as any;
 }
 
 export function resolveRaise<TContext, TEvent extends EventObject>(
@@ -200,7 +201,7 @@ export function resolveRaise<TContext, TEvent extends EventObject>(
     type: actionTypes.raise,
     _event: resolvedEvent,
     delay: resolvedDelay
-  };
+  } as any;
 }
 
 /**
@@ -236,7 +237,7 @@ export function send<
         : isFunction(event)
         ? event.name
         : (getEventType<TSentEvent>(event) as string)
-  };
+  } as any;
 }
 
 export function resolveSend<
@@ -282,7 +283,7 @@ export function resolveSend<
     _event: resolvedEvent,
     event: resolvedEvent.data,
     delay: resolvedDelay
-  };
+  } as any;
 }
 
 /**
@@ -397,22 +398,23 @@ export function log<TContext, TEvent extends EventObject>(
     type: actionTypes.log,
     label,
     expr
-  };
+  } as any;
 }
 
 export const resolveLog = <TContext, TEvent extends EventObject>(
   action: LogAction<TContext, TEvent>,
   ctx: TContext,
   _event: SCXML.Event<TEvent>
-): LogActionObject<TContext, TEvent> => ({
-  // TODO: remove .expr from resulting object
-  ...action,
-  value: isString(action.expr)
-    ? action.expr
-    : action.expr(ctx, _event.data, {
-        _event
-      })
-});
+): LogActionObject<TContext, TEvent> =>
+  ({
+    // TODO: remove .expr from resulting object
+    ...action,
+    value: isString(action.expr)
+      ? action.expr
+      : action.expr(ctx, _event.data, {
+          _event
+        })
+  } as any);
 
 /**
  * Cancels an in-flight `send(...)` action. A canceled sent action will not
@@ -421,11 +423,13 @@ export const resolveLog = <TContext, TEvent extends EventObject>(
  *
  * @param sendId The `id` of the `send(...)` action to cancel.
  */
-export const cancel = (sendId: string | number): CancelAction => {
+export const cancel = <TContext, TEvent extends EventObject>(
+  sendId: string | number
+): CancelAction<TContext, TEvent> => {
   return {
     type: actionTypes.cancel,
     sendId
-  };
+  } as any;
 };
 
 /**
@@ -442,7 +446,7 @@ export function start<TContext, TEvent extends EventObject>(
     type: ActionTypes.Start,
     activity: activityDef,
     exec: undefined
-  };
+  } as any;
 }
 
 /**
@@ -451,10 +455,7 @@ export function start<TContext, TEvent extends EventObject>(
  * @param actorRef The activity to stop.
  */
 export function stop<TContext, TEvent extends EventObject>(
-  actorRef:
-    | string
-    | ActivityDefinition<TContext, TEvent>
-    | Expr<TContext, TEvent, string | { id: string }>
+  actorRef: string | Expr<TContext, TEvent, string | { id: string }>
 ): StopAction<TContext, TEvent> {
   const activity = isFunction(actorRef)
     ? actorRef
@@ -464,7 +465,7 @@ export function stop<TContext, TEvent extends EventObject>(
     type: ActionTypes.Stop,
     activity,
     exec: undefined
-  };
+  } as any;
 }
 
 export function resolveStop<TContext, TEvent extends EventObject>(
@@ -499,7 +500,7 @@ export const assign = <TContext, TEvent extends EventObject = EventObject>(
   return {
     type: actionTypes.assign,
     assignment
-  };
+  } as any;
 };
 
 export function isActionObject<TContext, TEvent extends EventObject>(
@@ -575,14 +576,16 @@ export function pure<TContext, TEvent extends EventObject>(
     event: TEvent
   ) =>
     | SingleOrArray<
-        ActionObject<TContext, TEvent> | ActionObject<TContext, TEvent>['type']
+        | BaseActionObject
+        | BaseActionObject['type']
+        | ActionObject<TContext, TEvent>
       >
     | undefined
 ): PureAction<TContext, TEvent> {
   return {
     type: ActionTypes.Pure,
     get: getActions
-  };
+  } as any;
 }
 
 /**
@@ -653,7 +656,7 @@ export function choose<TContext, TEvent extends EventObject>(
   return {
     type: ActionTypes.Choose,
     conds
-  };
+  } as any;
 }
 
 const pluckAssigns = <TContext, TEvent extends EventObject>(
@@ -729,12 +732,14 @@ export function resolveActions<TContext, TEvent extends EventObject>(
         ) as SendActionObject<TContext, TEvent>; // TODO: fix ActionTypes.Init
 
         if (!IS_PRODUCTION) {
+          const configuredDelay = (
+            actionObject as SendAction<TContext, TEvent, AnyEventObject>
+          ).delay;
           // warn after resolving as we can create better contextual message here
           warn(
-            !isString(actionObject.delay) ||
-              typeof sendAction.delay === 'number',
+            !isString(configuredDelay) || typeof sendAction.delay === 'number',
             // tslint:disable-next-line:max-line-length
-            `No delay reference for delay expression '${actionObject.delay}' was found on machine '${machine.id}'`
+            `No delay reference for delay expression '${configuredDelay}' was found on machine '${machine.id}'`
           );
         }
 
@@ -776,35 +781,32 @@ export function resolveActions<TContext, TEvent extends EventObject>(
           return [];
         }
 
-        const [
-          resolvedActionsFromChoose,
-          resolvedContextFromChoose
-        ] = resolveActions(
-          machine,
-          currentState,
-          updatedContext,
-          _event,
-          [
-            {
-              type: blockType,
-              actions: toActionObjects(
-                toArray(matchedActions),
-                machine.options.actions as any
-              )
-            }
-          ],
-          predictableExec,
-          preserveActionOrder
-        );
+        const [resolvedActionsFromChoose, resolvedContextFromChoose] =
+          resolveActions(
+            machine,
+            currentState,
+            updatedContext,
+            _event,
+            [
+              {
+                type: blockType,
+                actions: toActionObjects(
+                  toArray(matchedActions),
+                  machine.options.actions as any
+                )
+              }
+            ],
+            predictableExec,
+            preserveActionOrder
+          );
         updatedContext = resolvedContextFromChoose;
         preservedContexts?.push(updatedContext);
         return resolvedActionsFromChoose;
       }
       case actionTypes.pure: {
-        const matchedActions = (actionObject as PureAction<
-          TContext,
-          TEvent
-        >).get(updatedContext, _event.data);
+        const matchedActions = (
+          actionObject as PureAction<TContext, TEvent>
+        ).get(updatedContext, _event.data);
         if (!matchedActions) {
           return [];
         }
@@ -834,7 +836,7 @@ export function resolveActions<TContext, TEvent extends EventObject>(
           actionObject as StopAction<TContext, TEvent>,
           updatedContext,
           _event
-        );
+        ) as any;
 
         predictableExec?.(resolved, currentContext, _event);
         return resolved;
@@ -859,12 +861,13 @@ export function resolveActions<TContext, TEvent extends EventObject>(
           predictableExec(resolvedActionObject, updatedContext, _event);
         } else if (exec && preservedContexts) {
           const contextIndex = preservedContexts.length - 1;
-          resolvedActionObject = {
+          const wrapped = {
             ...resolvedActionObject,
             exec: (_ctx, ...args) => {
-              exec(preservedContexts[contextIndex], ...args);
+              (exec as any)(preservedContexts[contextIndex], ...args);
             }
           };
+          resolvedActionObject = wrapped as any;
         }
         return resolvedActionObject;
     }