+++ packages/core/test/actions.test.ts
@@ -38,11 +38,11 @@ function trackEntries(machine: AnyStateMachine) {
     state.onEntry.unshift({
       type: '__testEntryTracker',
       exec: () => logs.push(`enter: ${stateDescription}`)
-    });
+    } as any);
     state.onExit.unshift({
       type: '__testExitTracker',
       exec: () => logs.push(`exit: ${stateDescription}`)
-    });
+    } as any);
   }
 
   function addTrackingActionsRecursively(
@@ -1792,7 +1792,7 @@ describe('actions config', () => {
           initialState.context,
           { type: 'any' },
           {
-            action,
+            action: action as any,
             state: initialState,
             _event: initialState._event
           }
@@ -1812,7 +1812,7 @@ describe('actions config', () => {
           inactiveState.context,
           { type: 'EVENT' },
           {
-            action,
+            action: action as any,
             state: initialState,
             _event: initialState._event
           }