+++ packages/core/test/invoke.test.ts
@@ -1859,7 +1859,9 @@ describe('invoke', () => {
         JSON.stringify(waitingState);
       }).not.toThrow();
 
-      expect(typeof waitingState.actions[0].activity!.src).toBe('string');
+      expect(typeof (waitingState.actions[0] as any).activity!.src).toBe(
+        'string'
+      );
     });
 
     it('should throw error if unhandled (sync)', () => {