+++ packages/xstate-react/test/useActor.test.tsx
@@ -11,9 +11,15 @@ import {
   spawn,
   toActorRef
 } from 'xstate';
-import { useMachine } from '../src';
+import { useInterpret, useMachine } from '../src';
 import { useActor } from '../src/useActor';
 import { describeEachReactMode } from './utils';
+
+const originalConsoleError = console.error;
+
+afterEach(() => {
+  console.error = originalConsoleError;
+});
 
 describeEachReactMode('useActor (%s)', ({ render }) => {
   it('initial invoked actor should be immediately available', (done) => {
@@ -504,5 +510,21 @@ describeEachReactMode('useActor (%s)', ({ render }) => {
     fireEvent.click(elSend);
 
     expect(elState.textContent).toEqual('two');
+  });
+
+  it('should not log any spurious errors when used with a not-started actor', () => {
+    const spy = jest.fn();
+    console.error = spy;
+
+    const machine = createMachine({});
+    const App = () => {
+      useActor(useInterpret(machine));
+
+      return null;
+    };
+
+    render(<App />);
+
+    expect(spy).not.toBeCalled();
   });
 });